+++ http/plugin.go
@@ -4,12 +4,10 @@ import (
 	"context"
 	"log"
 	"net/http"
-	"strconv"
 	"sync"
 	"time"
 
 	"github.com/mholt/acmez"
-	"github.com/prometheus/client_golang/prometheus"
 	endure "github.com/spiral/endure/pkg/container"
 	"github.com/spiral/errors"
 	"github.com/spiral/roadrunner-plugins/v2/config"
@@ -65,8 +63,7 @@ type Plugin struct {
 	eventsID string
 
 	// Pool which attached to all servers
-	pool        pool.Pool
-	writersPool sync.Pool
+	pool pool.Pool
 
 	// servers RR handler
 	handler *handler.Handler
@@ -120,13 +117,6 @@ func (p *Plugin) Init(cfg config.Configurer, rrLogger logger.Logger, server serv
 
 	p.server = server
 	p.events, p.eventsID = events.Bus()
-	p.writersPool = sync.Pool{
-		New: func() interface{} {
-			wr := new(writer)
-			wr.code = -1
-			return wr
-		},
-	}
 
 	return nil
 }
@@ -292,33 +282,20 @@ func (p *Plugin) Stop() error {
 
 // ServeHTTP handles connection using set of middleware and pool PSR-7 server.
 func (p *Plugin) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	tt := time.Now()
-	wr := p.getWriter(w)
-
 	defer func() {
 		err := r.Body.Close()
 		if err != nil {
 			p.log.Error("body close", "error", err)
 		}
-
-		requestCounter.With(prometheus.Labels{
-			"status": strconv.Itoa(wr.code),
-		}).Inc()
-
-		requestDuration.With(prometheus.Labels{
-			"status": strconv.Itoa(wr.code),
-		}).Observe(time.Since(tt).Seconds())
-
-		p.putWriter(wr)
 	}()
 
 	if headerContainsUpgrade(r) {
-		http.Error(wr, "server does not support upgrade header", http.StatusInternalServerError)
+		http.Error(w, "server does not support upgrade header", http.StatusInternalServerError)
 		return
 	}
 
 	if p.https != nil && r.TLS == nil && p.cfg.SSLConfig.Redirect {
-		p.redirect(wr, r)
+		p.redirect(w, r)
 		return
 	}
 
@@ -329,7 +306,7 @@ func (p *Plugin) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	r = attributes.Init(r)
 	// protect the case, when user sendEvent Reset, and we are replacing handler with pool
 	p.mu.RLock()
-	p.handler.ServeHTTP(wr, r)
+	p.handler.ServeHTTP(w, r)
 	p.mu.RUnlock()
 }